+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- Support for FullBleed configuration for adding web parts in "Full Width column" section [OliverZeiser]
 - Improvements to ExecuteQueryRetryAsync [OliverZeiser, biste5]
 - Improvements to support provisioning engine to be called from non console applications
 - Better support for async methods, avoiding deadlocks